:white_check_mark: [#3403] Theme admin tests update/additions
