+++ src/openforms/config/tests/factories.py
@@ -1,0 +1,8 @@
+import factory
+
+
+class ThemeFactory(factory.django.DjangoModelFactory):
+    name = factory.Faker("bs")
+
+    class Meta:
+        model = "config.Theme"
