+++ src/openforms/config/tests/test_global_configuration.py
@@ -1,7 +1,6 @@
 import json
 import shutil
 import tempfile
-from pathlib import Path
 from unittest.mock import patch
 
 from django.conf import settings
@@ -12,15 +11,11 @@ from django.utils.translation import gettext as _
 
 import clamd
 from django_webtest import WebTest
-from webtest import Upload
 
 from openforms.accounts.tests.factories import SuperUserFactory
-from openforms.forms.tests.factories import FormFactory
 from openforms.tests.utils import NOOP_CACHES, disable_2fa
 
 from ..models import GlobalConfiguration
-
-LOGO_FILE = Path(settings.BASE_DIR) / "docs" / "logo.svg"
 
 
 @disable_2fa
@@ -48,72 +43,6 @@ class AdminTests(WebTest):
                 pass
 
         self.addCleanup(_cleanup)
-
-    def test_upload_svg(self):
-        with self.subTest(part="admin config"):
-            url = reverse("admin:config_globalconfiguration_change", args=(1,))
-
-            change_page = self.app.get(url)
-
-            with open(LOGO_FILE, "rb") as infile:
-                upload = Upload("logo.svg", infile.read(), "image/svg+xml")
-
-            change_page.form["logo"] = upload
-            response = change_page.form.submit()
-
-            self.assertEqual(response.status_code, 302)
-            config = GlobalConfiguration.get_solo()
-            self.assertEqual(config.logo, "logo/logo.svg")
-
-        with self.subTest(part="logo used"):
-            form = FormFactory.create()
-            url = reverse("forms:form-detail", kwargs={"slug": form.slug})
-
-            form_page = self.app.get(url)
-
-            header = form_page.pyquery(".utrecht-page-header")
-            self.assertTrue(header)
-            self.assertIn(
-                "utrecht-page-header--openforms-with-logo", header.attr("class")
-            )
-
-            style_tag = form_page.pyquery("style")
-            self.assertIn(
-                f"--of-header-logo-url: url('{config.logo.url}')",
-                style_tag.text(),
-            )
-
-    def test_upload_png(self):
-        logo = Path(settings.DJANGO_PROJECT_DIR) / "static" / "img" / "digid.png"
-        url = reverse("admin:config_globalconfiguration_change", args=(1,))
-
-        change_page = self.app.get(url)
-
-        with open(logo, "rb") as infile:
-            upload = Upload("digid.png", infile.read(), "image/png")
-
-        change_page.form["logo"] = upload
-        response = change_page.form.submit()
-
-        self.assertEqual(response.status_code, 302)
-        config = GlobalConfiguration.get_solo()
-        self.assertEqual(config.logo, "logo/digid.png")
-
-    def test_upload_blank(self):
-        # fixes #581
-        config = GlobalConfiguration.get_solo()
-        config.logo = None
-        config.save()
-
-        url = reverse("admin:config_globalconfiguration_change", args=(1,))
-
-        change_page = self.app.get(url)
-
-        response = change_page.form.submit()
-
-        self.assertEqual(response.status_code, 302)
-        config = GlobalConfiguration.get_solo()
-        self.assertEqual(config.logo, "")
 
     def test_plugin_configuration(self):
         # mocking the admin/solo machinery is not straightforward here...
@@ -160,19 +89,6 @@ class AdminTests(WebTest):
         config.refresh_from_db()
         self.assertEqual(config.save_form_email_subject, "Subject {{form_name}}")
         self.assertEqual(config.save_form_email_content, "Content {{form_name}}")
-
-    def test_can_upload_stylesheet(self):
-        url = reverse("admin:config_globalconfiguration_change", args=(1,))
-        upload = Upload("my-theme.css", b".foo { display: block }", "text/css")
-
-        change_page = self.app.get(url)
-
-        change_page.form["theme_stylesheet_file"] = upload
-        response = change_page.form.submit()
-
-        self.assertEqual(response.status_code, 302)
-        config = GlobalConfiguration.get_solo()
-        self.assertEqual(config.theme_stylesheet_file, "config/themes/my-theme.css")
 
     @override_settings(LANGUAGE_CODE="en")
     def test_virus_scan_enabled_not_configured(self):
